refactor: move bardata to data
